finished the init migrations error

+++ src/infrastructure/persistence/db/models/menu.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 from uuid import UUID
 
 from clickhouse_sqlalchemy import types
@@ -14,7 +13,7 @@ class Menu(BaseCH):
     id: Mapped[UUID] = mapped_column(types.UUID, primary_key=True)
     name: Mapped[str] = mapped_column(types.String)
     description: Mapped[str] = mapped_column(types.String)
-    average_rating: Mapped[Decimal] = mapped_column(types.Decimal)
+    average_rating: Mapped[float] = mapped_column(types.Float)
     host_id: Mapped[UUID] = mapped_column(types.UUID)
 
     # TODO: sections, dinner_ids, menu_review_ids
